requirements.txt
numpy==1.18.5
Pillow==7.2.0
scipy==1.4.1
six==1.14.0
torch==1.4.0
torchvision==0.5.0
